+++ frontend/src/pages/LoginPage/index.tsx
@@ -1,13 +1,7 @@
-import { Link } from '@tanstack/react-router';
-
-import { Flex } from '@/components/Flex';
-import googleLoginIcon from '@/components/Icon/png/google-login-icon.png';
 import { Modal } from '@/components/Modal';
-import { Text } from '@/components/Text';
-import { serviceENV } from '@/envconfig';
-import { vars } from '@/theme/index.css';
+import { GoogleLoginButton } from '@/features/login/ui/GoogleLoginButton';
 
-import { backdropStyle, googleLoginButtonStyle } from './index.css';
+import { backdropStyle } from './index.css';
 
 const LoginPage = () => (
   <div className={backdropStyle}>
@@ -23,26 +17,6 @@ const LoginPage = () => (
       </Modal.Footer>
     </Modal>
   </div>
-);
-
-const GoogleLoginButton = () => (
-  <Flex
-    align='center'
-    as={Link}
-    className={googleLoginButtonStyle}
-    gap={200}
-    justify='center'
-    to={serviceENV.GOOGLE_OAUTH_URL}
-    width='full'
-  >
-    <img
-      alt='Google 로그인 아이콘'
-      height='20px'
-      src={googleLoginIcon}
-      width='20px'
-    />
-    <Text color={vars.color.Ref.Netural[400]} typo='t3'>Google로 시작하기</Text>
-  </Flex>
 );
 
 export default LoginPage;
